% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incrementalRMA.R
\name{cleanplatformname}
\alias{cleanplatformname}
\title{Clean platform name}
\usage{
cleanplatformname(abatch)
}
\arguments{
\item{abatch}{AffyBatch object to extract platform name from.}
}
\value{
A string representing a cleaned up platform name.
}
\description{
Return a cleaned up platform name from based on an
AffyBatch cdf name.
}
\details{
An AffyBatch records the CDF suitable for processing the
GeneChip data. Embedded in this cdf is the name of the
platform itself. This routine cleans up the cdf and
trims off the "cdf" on the platform name.

The \code{\link[affy]{AffyBatch-class} cdfName} function extracts a cdf
and the \code{\link[affy]{cleancdfname}} function cleans
up problems with the cdf name. However, this function goes
beyond this by stripping off the cdf text. This code
was taken from the \code{\link[frma]{frma}} package.
}
\examples{
\dontrun{
cleanplatformname(AffyBatch)
}

}
