% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMAError.R
\name{combine_and_deduplicate}
\alias{combine_and_deduplicate}
\title{Combine and De-duplicate AffyBatch objects.}
\usage{
combine_and_deduplicate(new, ref)
}
\arguments{
\item{new}{\code{\link[affy]{AffyBatch-class}} of samples to use}

\item{ref}{Reference \code{\link[affy]{AffyBatch-class}} to merge in non-duplicates for}
}
\value{
An \code{\link[affy]{AffyBatch-class}} combining the two arguments, with duplicates removed.
}
\description{
Having two AffyBatch objects, combine them into a single AffyBatch object, paying attention to duplicates.
}
\details{
Two AffyBatch objects (esets) can be combined using \code{combine} function. However, if they have
several samples in common we don't want the duplicates. This tries to remove duplicates by examining
both the names and the expression data for duplication.

The result of this effort is to produce a single AffyBatch which combines the two objects, but with
duplicates removed.
}
\examples{
\dontrun{
combine_and_deduplicate(Dilution, Dilution)
}
}
