% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incrementalRMA.R
\name{incrementalRMA}
\alias{incrementalRMA}
\title{Incremental RMA}
\usage{
incrementalRMA(abatch, params = NULL, calculate_error = FALSE)
}
\arguments{
\item{abatch}{An \code{\link[affy]{AffyBatch-class}} of cel files to
apply RMA to.}

\item{params}{An incremental parameter list as generated by
\code{\link{parameterizeRMA}}.}

\item{calculate_error}{incrementalRMA errors can be calculated vs. what canonical RMA
would produce.  See \code{\link{calculateIncrementalRMAError}} for a discussion
of the concept and interpretation.}
}
\value{
A \code{\link[Biobase]{ExpressionSet}} object representing the
RMA normalized AffyBatch object.
}
\description{
Perform incremental RMA algorithm on CEL files.
}
\details{
RMA is a robust multiarray average expression (see \code{\link[affy]{rma}}).
It is one of the de facto approaches for processing Affymetrix GeneChip
data.

One disadvantage of RMA is that parameters are calculated on the set of
arrays during the normalization and probeset summarization process. This
works out for studies in which all samples are available at once. In
the case of adding new samples into a dataset, the entire set must be
reprocessed with RMA. Making it even more challenging, this reprocessing
will result in slightly different parameters. These different parameters
lead in turn to slightly different gene expression measurements.

This package gains inspiration from the \code{frma} package, as well
as earlier efforts including refRMA and incrementalRMA. The goal of
the package is to use a set of data to pre-calculate necessary parameters
for RMA. Then, when a new sample is made available these stored parameters
can be applied to the new sample using the RMA approach. The result should
provide gene expression estimates similar to those used when the entire
dataset is used together.
}
\note{
A test of the package is the following: store the parameters from
normalizing a dataset D. Then apply incrementalRMA to each of the
raw samples in D. The resulting gene expression should be identical
in both cases.
}
\examples{
\dontrun{
incrementalRMA(AffyBatch, params=list(probeEffects=(),normalizationVector=()))
}
}
