% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incrementalRMA.R
\name{quantileNormalizeIncremental}
\alias{quantileNormalizeIncremental}
\title{Incremental quantile normalization}
\usage{
quantileNormalizeIncremental(abatch, params = NULL)
}
\arguments{
\item{abatch}{An AffyBatch to quantile normalize.}

\item{params}{An incremental parameter list as generated by
\code{\link{parameterizeRMA}}}
}
\value{
A matrix of pm (perfect match) values that have been normalized.
}
\description{
Use existing parameters to quantile normalize an AffyBatch.
}
\details{
An AffyBatch is normalized using a quantile normalization approach.
In the case of an incremental algorith, the pre-calculated quantiles
can be used to normalize new data against. This has the benefit of
not requiring calculation of quantiles across a set, instead just
applying a stored version.
}
\note{
The incremental quantile normalization information is stored
within the params object as the list \code{normalizationVector}.

If the params object is NULL then a default quantile normalization
is performed.
}
\examples{
\dontrun{
quantileNormalizeIncremental(AffyBatch, params=list("normalizationVector"=()))
}
}
